Corrections

+++ p2/mbroker/boxes.h
@@ -26,10 +26,10 @@ typedef struct {
 
 extern char err_msg[ERR_SIZE];
 extern box **system_boxes;
-extern long num_of_boxes;
+extern int num_of_boxes;
 
 void destroy_system_boxes();
-long find_box(char *box_name);
+int find_box(char *box_name);
 int remove_box(char *box_name);
 int comparator(const void *b1, const void *b2);
 void resize_system_boxes(box *new_box);
